+++ src/app/core/ase-layout/ase-layout.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+
+@Component({
+  selector: 'ase-layout',
+  standalone: true,
+  imports: [RouterOutlet],
+  templateUrl: './ase-layout.component.html',
+  styleUrl: './ase-layout.component.scss'
+})
+export class AseLayoutComponent {
+
+}