+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Add
 
 - Add project basic configuration
+- Add basic layout